+++ cogstat/cogstat_stat.py
@@ -202,7 +202,7 @@ def pivot(pdf, row_names, col_names, page_names, depend_name, function):
 
 
 def diffusion(df, error_name='', RT_name='', participant_name='', condition_names=[], correct_coding='0',
-              reaction_time_in='sec', scaling_parameter=0.1, case_unsensitive_index_sort=True):
+              reaction_time_in='sec', scaling_parameter=0.1, max_reaction_time=2.5, case_unsensitive_index_sort=True):
     """Drift diffusion parameter recovery based on the EZ method.
 
     Parameters
@@ -224,8 +224,10 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
         Unit of reaction time
     scaling_parameter : float
         Usually either 0.1 or 1
+    max_reaction_time : float
+        Maximum reaction time is sec. Trials with larger RT will be excluded.
     case_unsensitive_index_sort : bool
-        Pdf.pivot_table() sorts the index, but unlike spreadsheet software packages, it is case sensitive.
+        Pdf.pivot_table() sorts the index, but unlike spreadsheet software packages, it is case-sensitive.
         If this parameter is True, the indexes will be reordered to be case-insensitive
 
     Returns
@@ -233,19 +235,27 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
 
     """
 
+    # Can we run the analysis?
     if not (error_name and RT_name):
         result = _('Specify the minimum required parameters (reaction time and error).')
         return result
+
+    # Variable information
     result = ''
-    result += _('Error: %s, Reaction time: %s, Participant: %s, Condition(s): %s') % \
+    result += _('Used variables.') + ' ' + _('Error: %s, Reaction time: %s, Participant: %s, Condition(s): %s') % \
               (error_name, RT_name, participant_name if participant_name != '' else _('None'),
                ','.join(condition_names) if condition_names != [] else _('None'))
+
+    # 1. Prepare the raw data
+    # Use df_diff dataframe for the calculations
     df_diff = df.copy()
 
+    # Modify time unit and error coding if needed
     if reaction_time_in == 'msec':  # if time is in msec, then convert time to sec
         df_diff[RT_name] = df_diff[RT_name] / 1000
     if correct_coding == '1':  # if correct is coded as 1, then reverse the coding
         df_diff[error_name] = 1 - df_diff[error_name]
+
     # If condition and/or participant variables were not given, add a quasi condition/participant variable with
     # constant values
     if not condition_names:
@@ -257,8 +267,18 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
 
     # If any data is missing from a trial, drop the whole trial
     df_diff = df_diff.dropna(subset=[error_name] + [RT_name] + [participant_name] + condition_names)
+    missing_trials_n = len(df) - len(df_diff)
+    result += '\n\n' + _('Number of trials excluded because of missing data:') + \
+              ' %s (%.1f%%)\n' % (missing_trials_n, (missing_trials_n / len(df)) * 100)
 
-    # Calculate RT and error rate statistics
+    # Filter slow outlier cases
+    filtered_trials_n = len(df_diff) - len(df_diff[df_diff[RT_name] < max_reaction_time])
+    df_diff = df_diff[df_diff[RT_name] < max_reaction_time]
+    result += _('Number of trials excluded because reaction time was larger than %s sec:') % max_reaction_time + \
+              ' %s (%.1f%%)\n' % (filtered_trials_n, (filtered_trials_n / len(df)) * 100)
+
+    # 2. Calculate N, RT, and error rate statistics
+    n_table = pd.pivot_table(df_diff, values=error_name, index=participant_name, columns=condition_names, aggfunc=len)
     mean_correct_RT_table = pd.pivot_table(df_diff[df_diff[error_name] == 0], values=RT_name,
                                            index=participant_name, columns=condition_names, aggfunc=np.mean)
     var_correct_RT_table = pd.pivot_table(df_diff[df_diff[error_name] == 0], values=RT_name,
@@ -272,6 +292,7 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
         # we reorder the tables to be case-insensitive
         from pandas.api.types import is_string_dtype
         if is_string_dtype(mean_correct_RT_table.index):
+            n_table.sort_index(inplace=True, key=lambda x: x.str.lower())
             mean_correct_RT_table.sort_index(inplace=True, key=lambda x: x.str.lower())
             var_correct_RT_table.sort_index(inplace=True, key=lambda x: x.str.lower())
             mean_percent_correct_table.sort_index(inplace=True, key=lambda x: x.str.lower())
@@ -279,6 +300,7 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
     # Display RT and error rate statistics
     previous_precision = pd.get_option('display.precision')
     pd.set_option('display.precision', 3)  # thousandth in error, milliseconds in RT, thousandths in diffusion parameters
+    result += '\n\n' + _('Number of trials') + _format_html_table(n_table.to_html(bold_rows=False))
     result += '\n\n' + _('Mean percent correct with edge correction') + _format_html_table(mean_percent_correct_table.
                                                                                            to_html(bold_rows=False))
     result += '\n\n' + _('Mean correct reaction time') + _format_html_table(mean_correct_RT_table.
@@ -286,7 +308,7 @@ def diffusion(df, error_name='', RT_name='', participant_name='', condition_name
     result += '\n\n' + _('Correct reaction time variance') + _format_html_table(var_correct_RT_table.
                                                                                 to_html(bold_rows=False))
 
-    # Recover diffusion parameters
+    # 3. Recover diffusion parameters
     original_index = mean_percent_correct_table.index  # to recover index order later
     original_columns = mean_percent_correct_table.columns  # to recover column order later
     EZ_parameters = pd.concat([mean_percent_correct_table.stack(condition_names),